move shared functions back to clusterer.rs
